+++ src/ts/upload/index.ts
@@ -201,7 +201,6 @@ const uploadFiles = (vditor: IVditor, files: FileList | DataTransferItemList | F
 
     const extraData = vditor.options.upload.extraData;
     for (const key of Object.keys(extraData)) {
-        // @ts-ignore
         formData.append(key, extraData[key]);
     }