+++ src/ts/markdown/chartRender.ts
@@ -19,7 +19,7 @@ export const chartRender = (element: (HTMLElement | Document) = document,
                     echarts.init(e).setOption(option);
                     e.setAttribute("data-processed", "true");
                 } catch (error) {
-                    e.className = "hljs";
+                    e.className = "vditor-reset--error";
                     e.innerHTML = `echarts render error: <br>${error}`;
                 }
             });