+++ src/ts/markdown/highlightRender.ts
@@ -37,6 +37,7 @@ export const highlightRender = (hljsOption?: IHljs, element: HTMLElement | Docum
             }
 
             if (block.classList.contains("language-mermaid") || block.classList.contains("language-echarts")
+                || block.classList.contains("language-mindmap")
                 || block.classList.contains("language-abc") || block.classList.contains("language-graphviz")) {
                 return;
             }