+++ types/index.d.ts
@@ -152,6 +152,8 @@ interface ILute {
 
     FormatMd(markdown: string): string;
 
+    RenderMindmap(text: string): string;
+
     // debugger md
     RenderEChartsJSON(text: string): string;